refactor(core): share Rust operator read path
Move operator read-model loading and JSON projection builders into a shared Rust read_path module so status, summary, runs, explain, and compare commands use one projection boundary while keeping CLI formatting separate.
